chore(frontend): Improve slug for settings views
