+++ apps/frontend/app/[locale]/(main)/(vitnode)/[...slug]/page.tsx
@@ -1,6 +1,10 @@
-import { SlugView, SlugViewProps } from 'vitnode-frontend/views/slug';
+import {
+  SlugView,
+  SlugViewProps,
+  generateMetadataSlug,
+} from 'vitnode-frontend/views/slug';
 
-// export const generateMetadata = generateMetadataSlugAdmin;
+export const generateMetadata = generateMetadataSlug;
 
 export default function Page(props: SlugViewProps) {
   return <SlugView {...props} />;
