+++ apps/frontend/app/[locale]/(main)/(vitnode)/settings/[...slug]/page.tsx
@@ -1,0 +1,6 @@
+import { SlugViewProps } from 'vitnode-frontend/views/slug';
+import { SlugSettingsView } from 'vitnode-frontend/views/theme/views/settings/slug';
+
+export default function Page(props: SlugViewProps) {
+  return <SlugSettingsView {...props} />;
+}
