+++ apps/frontend/app/[locale]/admin/(auth)/(vitnode)/[...slug]/page.tsx
@@ -1,11 +1,11 @@
 import {
   SlugAdminView,
   generateMetadataSlugAdmin,
-  SlugAdminViewProps,
 } from 'vitnode-frontend/views/admin/views/slug';
+import { SlugViewProps } from 'vitnode-frontend/views/slug';
 
 export const generateMetadata = generateMetadataSlugAdmin;
 
-export default function Page(props: SlugAdminViewProps) {
+export default function Page(props: SlugViewProps) {
   return <SlugAdminView {...props} />;
 }
