+++ packages/frontend/src/views/theme/views/settings/nav/hooks/use-settings-view.ts
@@ -24,7 +24,7 @@ export const useSettingsView = (): ReturnValues => {
   return {
     navItems: [
       {
-        href: '/settings',
+        href: '/settings/overview',
         icon: Cog,
         text: t('overview.title'),
       },
